using travis ci
- モデルをモジュールに
- travis.ymlの追加
- travis上でgae sdkを使用する
