+++ .travis.yml
@@ -1,0 +1,23 @@
+language: python
+python:
+  - '2.7'
+before_install:
+  - openssl aes-256-cbc -K $encrypted_4614d7b9ffa9_key -iv $encrypted_4614d7b9ffa9_iv -in pkmn-tool-secret.json.enc -out pkmn-tool-secret.json -d
+  - pushd $HOME
+  - python $TRAVIS_BUILD_DIR/ciscripts/getlatestsdk.py
+  - unzip -q python_appengine_sdk-latest.zip
+  - cd google_appengine
+  - export GAE=$(pwd)
+  - export PYTHONPATH=$GAE:$PYTHONPATH
+  - export PYTHONPATH=$GAE/lib/yaml/lib:$PYTHONPATH
+  - export PYTHONPATH=$TRAVIS_BUILD_DIR/lib:$PYTHONPATH
+  - popd
+install:
+  - pip install -r requirements.txt
+  - pip install -r requirements-vendor.txt -t lib/
+  - env
+script: py.test front/tests/*
+deploy:
+  provider: gae
+  keyfile: pkmn-tool-secret.json.enc
+  project: pkmn-tool
