+++ front/models/__init__.py
@@ -1,0 +1,2 @@
+from ._type import Type, TYPES
+from ._pokemon import Pokemon, MAX_POKEMONS, MAX_POKEMONS_NO
