+++ mysite/settings.py
@@ -39,7 +39,7 @@ SECRET_KEY = os.getenv('DJANGO_SECRET_KEY', '')
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = False
 
-ALLOWED_HOSTS = ['localhost', '127.0.0.1', '*appspot.com*']
+ALLOWED_HOSTS = ['localhost', '127.0.0.1', '.appspot.com']
 
 
 # Application definition
