+++ requirements.txt
@@ -1,1 +1,2 @@
+mock==2.0.0
 pytest==2.6.4
